refactor: generic ws server to take ports
