+++ src-tauri/src/main.rs
@@ -7,9 +7,12 @@ mod services;
 use anyhow::Context;
 
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, EnvFilter, Registry};
-use crate::services::{run_auth_server, start_obs_server};
+use crate::services::{run_auth_server, start_ws_server};
 
 fn main() -> anyhow::Result<()> {
+    let obs_port = "37891";
+    let streamdeck_port = "17448";
+    
     Registry::default()
         .with(if cfg!(debug_assertions) {
             EnvFilter::new("info,tts_helper=trace")
@@ -27,8 +30,8 @@ fn main() -> anyhow::Result<()> {
             // Run auth server
             let handle = app.handle();
             std::thread::spawn(move || run_auth_server(handle));
-            tauri::async_runtime::spawn(start_obs_server());
-            //std::thread::spawn(move || run_websocket_server());
+            tauri::async_runtime::spawn(start_ws_server(obs_port));
+            tauri::async_runtime::spawn(start_ws_server(streamdeck_port));
 
             Ok(())
         })
