+++ src-tauri/src/services.rs
@@ -1,7 +1,5 @@
 mod oauth_listener;
 mod websocket_server;
-mod obs_websocket;
 
 pub use oauth_listener::*;
 pub use websocket_server::*;
-pub use obs_websocket::*;
